migrate psha_adapter functions out of cli script;
